Change value sync method
Rather than using late updates and relying on the client to send the value, use the fill/drain started and stopped events. set a flag when filling/draining starts and then on server tick, communicate the latest fill value for all resources filling or draining.

Upon release clear the flag and send the final value.

Server side tracking of which resources are being interacted with is required for updating players entering the culling zone